+++ src/EUITabPanel.cpp
@@ -3,7 +3,7 @@
 #include "EUITabSheet.h"
 #include "native/win/WinTabPanel.h"
 
-EUITabPanel::EUITabPanel(EUIWidget* prnt, float set_x, float set_y, float set_w, float set_h) : EUIWidget(prnt, "")
+EUITabPanel::EUITabPanel(EUIWidget* prnt, int set_x, int set_y, int set_w, int set_h) : EUIWidget(prnt, "")
 {
 	x = set_x;
 	y = set_y;
